backend/user_auth_microservice/requirements.txt
grpcio
grpcio-tools
PyJWT
peewee
mysqlclient
bcrypt
aio-pika
redis
python-dotenv
fastapi
uvicorn[standard]